Initial command split
