+++ cmd/main.go
@@ -1,0 +1,12 @@
+package main
+
+import (
+	"github.com/instrumenta/conftest/pkg/commands"
+	"os"
+)
+
+func main() {
+	if err := commands.NewDefaultCommand().Execute(); err != nil {
+		os.Exit(1)
+	}
+}
